requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
cattrs==25.3.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-filter==25.2
django-restframework==0.0.1
djangorestframework==3.16.1
fastf1==3.6.1
fonttools==4.60.1
idna==3.11
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
RapidFuzz==3.14.3
redis==5.2.1
requests==2.32.5
requests-cache==1.2.1
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
timple==0.1.8
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==13.1
